# perf: time.format_rfc3339 and time.now() [10x faster together]