### TODO
- [ ] replace half by `now(presicion ?string)` // presicion will be `nanosecond`, `milisecond`, `sec` for performance purpose

### Benchmark
```v
import time
import benchmark
import os

const max_iterations = os.getenv_opt('MAX_ITERATIONS') or { '100000' }.int()

fn main() {
	empty_time := time.Time{}
	// now_time := time.now()
	// fast_now_time := time.fast_linux_now()


	mut b := benchmark.start()

	for _ in 0 .. max_iterations {
		_ := empty_time.fast_format_rfc3339()
	}

	b.measure('empty_time.fast_format_rfc3339()')

	for _ in 0 .. max_iterations {
		_ := empty_time.format_rfc3339()
	}

	b.measure('empty_time.format_rfc3339()\n')
// -------------------------------------------------------------------------------
	for _ in 0 .. max_iterations {
		_ := time.now().fast_format_rfc3339()
	}

	b.measure('time.now().fast_format_rfc3339()')


	for _ in 0 .. max_iterations {
		_ := time.now().format_rfc3339()
	}

	b.measure('time.now().format_rfc3339()\n')
// -------------------------------------------------------------------------------
	for _ in 0 .. max_iterations {
		_ := time.fast_linux_now().fast_format_rfc3339()
	}

	b.measure('time.fast_linux_now().fast_format_rfc3339()')


	for _ in 0 .. max_iterations {
		_ := time.fast_linux_now().format_rfc3339()
	}

	b.measure('time.fast_linux_now().format_rfc3339()\n')
// -------------------------------------------------------------------------------
}
```
```bash
╰─ ./v -prod crun la.v ─╯
 SPENT    20.646 ms in empty_time.fast_format_rfc3339()
 SPENT     0.350 ms in empty_time.format_rfc3339()

 SPENT   140.616 ms in time.now().fast_format_rfc3339()
 SPENT   270.342 ms in time.now().format_rfc3339() // [Worse]

 SPENT    20.827 ms in time.fast_linux_now().fast_format_rfc3339() // [Better]
 SPENT   152.914 ms in time.fast_linux_now().format_rfc3339()

```

### time benchs
```v
import benchmark

const max_iterations = 100_000

fn main() {
	mut b := benchmark.start()
	mut ts := C.timespec{}

	for _ in 0 .. max_iterations {
		C.clock_gettime(C.CLOCK_REALTIME, &ts)
	}

	b.measure('CLOCK_REALTIME')

	for _ in 0 .. max_iterations {
		C.clock_gettime(C.CLOCK_REALTIME_COARSE, &ts)
	}

	b.measure('CLOCK_REALTIME_COARSE')

	for _ in 0 .. max_iterations {
		C.clock_gettime(C.CLOCK_MONOTONIC_COARSE, &ts)
	}

	b.measure('CLOCK_MONOTONIC_COARSE')

	for _ in 0 .. max_iterations {
		C.clock_gettime(C.CLOCK_MONOTONIC, &ts)
	}

	b.measure('CLOCK_MONOTONIC')

	for _ in 0 .. max_iterations {
		C.clock_gettime(C.CLOCK_MONOTONIC_RAW, &ts)
	}

	b.measure('CLOCK_MONOTONIC_RAW')

	for _ in 0 .. max_iterations {
		C.clock_gettime(C.CLOCK_BOOTTIME, &ts)
	}

	b.measure('CLOCK_BOOTTIME')

	for _ in 0 .. max_iterations {
		C.clock_gettime(C.CLOCK_PROCESS_CPUTIME_ID, &ts)
	}

	b.measure('CLOCK_PROCESS_CPUTIME_ID')

	for _ in 0 .. max_iterations {
		C.clock_gettime(C.CLOCK_THREAD_CPUTIME_ID, &ts)
	}

	b.measure('CLOCK_THREAD_CPUTIME_ID')

}
```
```bash
╰─ v -prod crun time_perf_teste.v ─╯
 SPENT   117.176 ms in CLOCK_REALTIME
 SPENT     0.546 ms in CLOCK_REALTIME_COARSE
 SPENT     0.523 ms in CLOCK_MONOTONIC_COARSE
 SPENT   118.127 ms in CLOCK_MONOTONIC
 SPENT   117.972 ms in CLOCK_MONOTONIC_RAW
 SPENT   119.232 ms in CLOCK_BOOTTIME
 SPENT    43.921 ms in CLOCK_PROCESS_CPUTIME_ID
 SPENT    40.514 ms in CLOCK_THREAD_CPUTIME_ID
```