@enghitalo 2023-12-31
> If you look at the `format_rfc3339` function, the first thing it calls is `local_to_utc`. That function looks like this:
> 
> ```v
> // local_to_utc converts the receiver `t` to the corresponding UTC time, if it contains local time.
> // If the receiver already does contain UTC time, it returns it unchanged.
> pub fn (t Time) local_to_utc() Time {
> 	if !t.is_local {
> 		return t
> 	}
> 	return Time{
> 		...t.add(-offset() * time.second)
> 		is_local: false
> 	}
> }
> ```
> 
> Since an empty Time struct won't have `is_local` set, it will always return quicker than a correctly set up Time struct. That will definitely cause a difference in the formatting time needed.
> 
> Also, I'm not seeing a huge difference like you are. Using latest V at this time, `V 0.4.3 9694024`, I am seeing this:
> 
> ```
>  SPENT    15.383 ms in empty_time
>  SPENT    24.246 ms in now_time
> ```
> 
> I have not tried applying your changes, this is what I see with current V.

It is, problably a memory effect, try change something on your code (like write a space randomly) and run again