@enghitalo 2023-12-31
> ```v
> import time
> import benchmark
> import os
> 
> const max_iterations = os.getenv_opt('MAX_ITERATIONS') or { '100000' }.int()
> 
> fn main() {
> 	empty_time := time.Time{}
> 	now_time := time.now()
> 
> 
> 	mut b := benchmark.start() 
> 
> 	for _ in 0 .. max_iterations {
> 		_ := empty_time.format_rfc3339()
> 	}
> 
> 	b.measure('empty_time')
> 
> 	for _ in 0 .. max_iterations {
> 		_ := now_time.format_rfc3339()
> 	}
> 
> 	b.measure('now_time')
> }
> ```

What is your OS?